+++ workspaces/rbac/plugins/rbac/src/components/ConditionalAccess/AddNestedConditionButton.tsx
@@ -15,28 +15,30 @@
  */
 import React from 'react';
 
-import { Box, Typography } from '@material-ui/core';
-import Tooltip from '@material-ui/core/Tooltip';
-import HelpOutlineIcon from '@material-ui/icons/HelpOutline';
+import HelpOutlineIcon from '@mui/icons-material/HelpOutline';
+import Box from '@mui/material/Box';
+import Tooltip from '@mui/material/Tooltip';
+import Typography from '@mui/material/Typography';
+
+export const tooltipTitle = () => (
+  <div>
+    <Typography variant="body1" component="p" align="center">
+      Nested conditions are <b>1 layer rules within a main condition</b>. It
+      lets you allow appropriate access by using detailed permissions based on
+      various conditions. You can add multiple nested conditions.
+    </Typography>
+    <Typography variant="body1" component="p" align="center">
+      For example, you can allow access to all entity types in the main
+      condition and use a nested condition to limit the access to entities owned
+      by the user.
+    </Typography>
+  </div>
+);
 
 export const AddNestedConditionButton = () => {
-  const tooltipTitle = () => (
-    <div>
-      <Typography variant="body1" component="p" align="center">
-        Nested conditions are <b>1 layer rules within a main condition</b>. It
-        lets you allow appropriate access by using detailed permissions based on
-        various conditions. You can add multiple nested conditions.
-      </Typography>
-      <Typography variant="body1" component="p" align="center">
-        For example, you can allow access to all entity types in the main
-        condition and use a nested condition to limit the access to entities
-        owned by the user.
-      </Typography>
-    </div>
-  );
   return (
     <Box
-      style={{
+      sx={{
         display: 'flex',
         justifyContent: 'center',
         alignItems: 'center',