+++ workspaces/rbac/plugins/rbac/src/components/CreateRole/AddedMembersTableColumn.tsx
@@ -18,8 +18,9 @@ import React from 'react';
 import { parseEntityRef } from '@backstage/catalog-model';
 import { Link, TableColumn } from '@backstage/core-components';
 
-import { Box, IconButton } from '@material-ui/core';
 import Delete from '@mui/icons-material/Delete';
+import Box from '@mui/material/Box';
+import IconButton from '@mui/material/IconButton';
 import { FormikErrors } from 'formik';
 
 import { RoleFormValues, SelectedMember } from './types';
@@ -98,6 +99,8 @@ export const selectedMembersColumns = (
               onClick={() => onRemove(mem.etag)}
               aria-label="Remove"
               title="Remove member"
+              style={{ padding: '0.5rem', borderRadius: '50%' }}
+              sx={{ '&:hover': { borderRadius: '50%' } }}
             >
               <Delete />
             </IconButton>