+++ workspaces/rbac/plugins/rbac/src/components/CreateRole/PoliciesCheckboxGroup.tsx
@@ -15,11 +15,11 @@
  */
 import React from 'react';
 
-import Checkbox from '@material-ui/core/Checkbox';
-import FormControl from '@material-ui/core/FormControl';
-import FormControlLabel from '@material-ui/core/FormControlLabel';
-import FormGroup from '@material-ui/core/FormGroup';
-import FormLabel from '@material-ui/core/FormLabel';
+import Checkbox from '@mui/material/Checkbox';
+import FormControl from '@mui/material/FormControl';
+import FormControlLabel from '@mui/material/FormControlLabel';
+import FormGroup from '@mui/material/FormGroup';
+import FormLabel from '@mui/material/FormLabel';
 
 import { PermissionsData } from '../../types';
 import { RowPolicy } from './types';