+++ workspaces/rbac/plugins/rbac/src/components/RoleOverview/MembersCard.tsx
@@ -19,13 +19,15 @@ import { parseEntityRef } from '@backstage/catalog-model';
 import { Table, WarningPanel } from '@backstage/core-components';
 import { usePermission } from '@backstage/plugin-permission-react';
 
-import { Card, CardContent, makeStyles } from '@material-ui/core';
-import CachedIcon from '@material-ui/icons/Cached';
+import CachedIcon from '@mui/icons-material/Cached';
+import Box from '@mui/material/Box';
+import Card from '@mui/material/Card';
+import CardContent from '@mui/material/CardContent';
 
 import { policyEntityUpdatePermission } from '@backstage-community/plugin-rbac-common';
 
 import { MembersInfo } from '../../hooks/useMembers';
-import { MembersData } from '../../types';
+import { filterTableData } from '../../utils/filter-table-data';
 import { getMembers } from '../../utils/rbac-utils';
 import EditRole from '../EditRole';
 import { columns } from './MembersListColumns';
@@ -34,14 +36,6 @@ type MembersCardProps = {
   roleName: string;
   membersInfo: MembersInfo;
 };
-
-const useStyles = makeStyles(theme => ({
-  empty: {
-    padding: theme.spacing(2),
-    display: 'flex',
-    justifyContent: 'center',
-  },
-}));
 
 const getRefreshIcon = () => <CachedIcon />;
 const getEditIcon = (isAllowed: boolean, roleName: string) => {
@@ -59,13 +53,12 @@ const getEditIcon = (isAllowed: boolean, roleName: string) => {
 
 export const MembersCard = ({ roleName, membersInfo }: MembersCardProps) => {
   const { data, loading, retry, error, canReadUsersAndGroups } = membersInfo;
-  const [members, setMembers] = React.useState<MembersData[]>();
+  const [searchText, setSearchText] = React.useState<string>();
   const policyEntityPermissionResult = usePermission({
     permission: policyEntityUpdatePermission,
     resourceRef: policyEntityUpdatePermission.resourceType,
   });
 
-  const classes = useStyles();
   const actions = [
     {
       icon: getRefreshIcon,
@@ -91,9 +84,10 @@ export const MembersCard = ({ roleName, membersInfo }: MembersCardProps) => {
     },
   ];
 
-  const onSearchResultsChange = (searchResults: MembersData[]) => {
-    setMembers(searchResults);
-  };
+  const filteredData = React.useMemo(
+    () => filterTableData({ data, columns, searchText }),
+    [data, searchText],
+  );
 
   return (
     <Card>
@@ -110,20 +104,23 @@ export const MembersCard = ({ roleName, membersInfo }: MembersCardProps) => {
         <Table
           title={
             !loading && data?.length
-              ? `Users and groups (${getMembers(members || data)})`
+              ? `Users and groups (${getMembers(filteredData)})`
               : 'Users and groups'
           }
           actions={actions}
-          renderSummaryRow={summary => onSearchResultsChange(summary.data)}
           options={{ padding: 'default', search: true, paging: true }}
           data={data ?? []}
           isLoading={loading}
           columns={columns}
           emptyContent={
-            <div data-testid="members-table-empty" className={classes.empty}>
+            <Box
+              data-testid="members-table-empty"
+              sx={{ display: 'flex', justifyContent: 'center', p: 2 }}
+            >
               No records found
-            </div>
+            </Box>
           }
+          onSearchChange={setSearchText}
         />
       </CardContent>
     </Card>