+++ workspaces/sentry/plugins/sentry/src/alpha/entityCard.tsx
@@ -16,6 +16,7 @@
 import React from 'react';
 import { compatWrapper } from '@backstage/core-compat-api';
 import { EntityCardBlueprint } from '@backstage/plugin-catalog-react/alpha';
+import { isSentryAvailable } from '../components';
 
 /**
  * @alpha
@@ -23,7 +24,7 @@ import { EntityCardBlueprint } from '@backstage/plugin-catalog-react/alpha';
 export const entitySentryCard = EntityCardBlueprint.make({
   name: 'sentry-issues',
   params: {
-    filter: 'kind:component',
+    filter: isSentryAvailable,
     loader: () =>
       import('../components/SentryIssuesWidget').then(m =>
         compatWrapper(<m.SentryIssuesWidgetCard />),