+++ workspaces/sonarqube/plugins/sonarqube/src/alpha/entityContent.tsx
@@ -15,6 +15,7 @@
  */
 import React from 'react';
 import { EntityContentBlueprint } from '@backstage/plugin-catalog-react/alpha';
+import { isSonarQubeAvailable } from '@backstage-community/plugin-sonarqube-react';
 
 /**
  * @alpha
@@ -24,7 +25,7 @@ export const entitySonarQubeContent = EntityContentBlueprint.make({
   params: {
     defaultPath: 'sonarqube',
     defaultTitle: 'SonarQube',
-    filter: 'kind:component',
+    filter: isSonarQubeAvailable,
     loader: () =>
       import('../components/SonarQubeContentPage').then(m => (
         <m.SonarQubeContentPage />