+++ workspaces/entity-feedback/plugins/entity-feedback-backend/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @backstage-community/plugin-entity-feedback-backend
 
+## 0.4.0
+
+### Minor Changes
+
+- fcf6f84: **BREAKING** `auth`, `config`, and `httpAuth` are now required, please migrate to the new backend system as the best path forward for this change.
+
+  Also, removed usages and references of `@backstage/backend-common`
+
 ## 0.3.2
 
 ### Patch Changes