+++ workspaces/jenkins/plugins/jenkins/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @backstage-community/plugin-jenkins
 
+## 0.12.2
+
+### Patch Changes
+
+- 69f11c0: Fix typo in the readme
+
+## 0.12.1
+
+### Patch Changes
+
+- e657f8d: Add default filtering based on required annotation to the new frontend system components
+
+## 0.12.0
+
+### Minor Changes
+
+- 6bde043: Add title property to Jenkins components.
+
 ## 0.11.1
 
 ### Patch Changes