+++ workspaces/ocm/plugins/ocm-backend/CHANGELOG.md
@@ -1,5 +1,20 @@
 ### Dependencies
 
+## 5.2.4
+
+### Patch Changes
+
+- 056a5ca: Updated dependency `@openapitools/openapi-generator-cli` to `2.15.3`.
+- 40f3226: Clean up api report warnings and remove unnecessary files
+- Updated dependencies [40f3226]
+  - @backstage-community/plugin-ocm-common@3.6.2
+
+## 5.2.3
+
+### Patch Changes
+
+- a6e850f: Updated dependency `msw` to `1.3.5`.
+
 ## 5.2.2
 
 ### Patch Changes