+++ workspaces/ocm/plugins/ocm-common/CHANGELOG.md
@@ -1,5 +1,11 @@
 ## @backstage-community/plugin-ocm-common [3.3.0](https://github.com/janus-idp/backstage-plugins/compare/@backstage-community/plugin-ocm-common@3.2.0...@backstage-community/plugin-ocm-common@3.3.0) (2024-07-26)
 
+## 3.6.2
+
+### Patch Changes
+
+- 40f3226: Clean up api report warnings and remove unnecessary files
+
 ## 3.6.1
 
 ### Patch Changes