+++ workspaces/pingidentity/plugins/catalog-backend-module-pingidentity/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @backstage-community/plugin-catalog-backend-module-pingidentity
 
+## 0.1.6
+
+### Patch Changes
+
+- a6e850f: Updated dependency `msw` to `1.3.5`.
+
 ## 0.1.5
 
 ### Patch Changes