+++ workspaces/quay/plugins/quay/CHANGELOG.md
@@ -1,5 +1,17 @@
 ### Dependencies
 
+## 1.14.5
+
+### Patch Changes
+
+- 41a3d82: Updated dependency `@playwright/test` to `1.48.2`.
+
+## 1.14.4
+
+### Patch Changes
+
+- a6e850f: Updated dependency `msw` to `1.3.5`.
+
 ## 1.14.3
 
 ### Patch Changes